Moved auth feature tests to their own folder/namespace and updated home page to dashboard
