+++ tests/Feature/Auth/DeleteAccountTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Auth;
 
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
