+++ tests/Feature/Auth/EmailVerificationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Auth;
 
 use App\Models\User;
 use Illuminate\Auth\Events\Verified;
